Changed location of addContact & addGroup buttons
update GUI
