+++ Chat-Application-Client/src/main/java/gov/iti/jets/presentation/controllers/ChatScreenController.java
@@ -1,30 +1,16 @@
 package gov.iti.jets.presentation.controllers;
 
-import java.io.IOException;
 import java.net.URL;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.ResourceBundle;
-import java.util.stream.Collector;
-import java.util.stream.Collectors;
-
-import javafx.collections.FXCollections;
-import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.fxml.Initializable;
 import javafx.scene.Parent;
-import javafx.scene.Scene;
-import javafx.scene.control.ListCell;
-import javafx.scene.control.ListView;
-import javafx.scene.control.TextField;
 import javafx.scene.control.Tooltip;
 import javafx.scene.image.ImageView;
 import javafx.scene.input.MouseEvent;
 import javafx.scene.layout.GridPane;
-import javafx.stage.Modality;
-import javafx.stage.Stage;
-import javafx.util.Callback;
+
 
 public class ChatScreenController implements Initializable {
 
@@ -47,18 +33,10 @@ public class ChatScreenController implements Initializable {
     GridPane mainGrid;
 
     @FXML
-    private ImageView addContactBtn;
+    private ImageView notificationsBtn;
 
     @FXML
-    private ImageView addgroupBtn;
-
-    @FXML
-    private ImageView notifications;
-
-    @FXML
-    private ImageView invitations;
-
-    private Parent root;
+    private ImageView invitationsBtn;
 
     @FXML
     @Override
@@ -68,8 +46,8 @@ public class ChatScreenController implements Initializable {
         Tooltip.install(btnGroups, new Tooltip("Groups"));
         Tooltip.install(btnContacts, new Tooltip("Contacts"));
         Tooltip.install(btnSettings, new Tooltip("Settings"));
-        Tooltip.install(addContactBtn, new Tooltip("Invite Contact"));
-        Tooltip.install(addgroupBtn, new Tooltip("Create Group"));
+        Tooltip.install(notificationsBtn, new Tooltip("Notifications"));
+        Tooltip.install(invitationsBtn, new Tooltip("Invitations"));
 
         openContacts();
 
@@ -86,49 +64,18 @@ public class ChatScreenController implements Initializable {
         }
 
     }
-
+    
     @FXML
-    private void clickAddContactBtn(MouseEvent event) {
-        InviteContactController inviteCont = null;
+    private void clkGroupBtn(MouseEvent event) {
+
         try {
-            FXMLLoader fxmlLoader = new FXMLLoader(getClass().getResource("/views/inviteContact.fxml"));
-            root = fxmlLoader.load();
-            inviteCont = fxmlLoader.getController();
-        } catch (IOException e) {
+            Parent f = (Parent) FXMLLoader.load(getClass().getResource("/views/groups.fxml"));
+            mainGrid.add(f, 1, 0);
+        } catch (Exception e) {
             e.printStackTrace();
         }
-        Stage stage = new Stage();
-        stage.initModality(Modality.APPLICATION_MODAL);
-        stage.setTitle("Invite Contacts");
 
-        inviteCont.setStage(stage);
-
-        Scene scene1 = new Scene(root, 501, 345);
-
-        stage.setScene(scene1);
-        stage.setResizable(false);
-        stage.showAndWait();
-    }
-
-    @FXML
-    private void clickAddGroupBtn(MouseEvent event) {
-        CreateGroupController createGroupCont = null;
-        try {
-            FXMLLoader fxmlLoader = new FXMLLoader(getClass().getResource("/views/createGroup.fxml"));
-            root = fxmlLoader.load();
-            createGroupCont = fxmlLoader.getController();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-        Stage stage = new Stage();
-        stage.initModality(Modality.APPLICATION_MODAL);
-        stage.setTitle("Create Group");
-        createGroupCont.setStage(stage);
-        Scene scene1 = new Scene(root, 501, 400);
-        stage.setScene(scene1);
-        stage.setResizable(false);
-        stage.showAndWait();
-    }
+    }  
 
     @FXML
     private void clkNotifications(MouseEvent event) {
